Bug fixing & processor work.
- Fixed bug in db:populate where some keys were added with a trailing '\n' char.
- Implemented Clone & Pull in the project model.
- Added classes to be Localization Processors responsible for each different project type.

+++ app/models/project.rb
@@ -1,7 +1,9 @@
 require 'logger'
 require 'rubygems'
 require 'git-ssh-wrapper'
+require 'find'
 require_dependency 'extensions/dir_extensions'
+require_dependency 'loc_processors/processor_factory'
 
 class Project < ActiveRecord::Base
     attr_accessible :name, :default_language_id, :project_type_id, :repository_address, :repository_ssh_key, :repository_type_id, :user_ids, :cloned
@@ -16,43 +18,60 @@ class Project < ActiveRecord::Base
 
     before_save :populate_users, :if => :users_changed?
 
-    def repository_pull 
-        unless self.cloned
-            self.git_clone
-            return 
+    def repository_clone
+        # Make sure that the directory exists.
+        Dir.mkpath self.get_repository_path
+
+        # Clone the repository.
+        wrapper = GitSSHWrapper.new(:private_key => self.repository_ssh_key, :log_level => 'ERROR')
+        logger.debug `env #{wrapper.cmd_prefix} git clone #{self.repository_address} #{self.get_repository_path}`
+        
+        # Update the project to indicate that it was cloned.
+        self.cloned = true;
+        self.save
+    ensure
+        wrapper.unlink if wrapper
+    end
+
+    def repository_pull
+        # Make sure that our project was cloned a first time.
+        if !self.cloned
+            return
         end
-        l = Logger.new('log/debug.log')
-        l.level = Logger::DEBUG
-        l.debug("in pull"+self.defaultLanguage.name)
+
+        # Pull the repository.
+        wrapper = GitSSHWrapper.new(:private_key => self.repository_ssh_key, :log_level => 'ERROR')
+        logger.debug `cd #{self.get_repository_path} && env #{wrapper.cmd_prefix} git pull `
+    ensure
+        wrapper.unlink if wrapper
     end
 
     def repository_push
 
     end
 
-    def repository_clone
-        # Make sure that the directory exists.
-        repository_path = File.join Rails.root, "repositories", self.id.to_s
-        Dir.mkpath repository_path
+    def repository_scan
+        # Make sure that our project was cloned a first time.
+        if !self.cloned
+            return
+        end
 
-        # Clone the repository.
-        wrapper = GitSSHWrapper.new(:private_key_path => '~/.ssh/id_rsa', :log_level => 'ERROR')
-        @hey = `env #{wrapper.cmd_prefix} git clone https://github.com/quentez/Klaim-iOS.git #{repository_path}`
+        # Retrieve the Loc Processor for our Project Type.
+        loc_processor = ProcessorFactory.get_processor self.project_type
 
-        logger.debug @hey
-        
-        # Update the project to indicate that it was cloned.
-        self.cloned = true;
-        self.save
-    ensure
-        wrapper.unlink
+        # Use the Loc Processor to find the Loc Files in the repository.
+        loc_processor.find_files Find.find(self.get_repository_path), self.default_language.format
     end
         
     def users_changed?
         self.user_ids && self.user_ids.size > 1
     end
 
-    private
+    protected
+
+    def get_repository_path
+        File.join Rails.root, "repositories", self.id.to_s
+    end
 
     def populate_users
         self.user_ids.delete ""
